Remove "_model" from Model filenames and rename BaseModel to Model and to __model__.py with cleanup
