+++ codes/models/__init__.py
@@ -1,38 +1,33 @@
 import logging
 
-logger = logging.getLogger("base")
+log = logging.getLogger("base")
 
 
 def create_model(opt):
     model = opt["model"]
 
     if model == "sr":
-        from .SR_model import SRModel as M
+        from .SR import SRModel as Md
     # elif model == 'srgan':
-    # from .SRGAN_model import SRGANModel as M
-    elif (
-            model == "srgan"
-            or model == "srragan"
-            or model == "srragan_hfen"
-            or model == "lpips"
-    ):
-        from .SRRaGAN_model import SRRaGANModel as M
+    # from .SRGAN_model import SRGANModel as Md
+    elif model == "srgan" or model == "srragan" or model == "srragan_hfen" or model == "lpips":
+        from .SRRaGAN import SRRaGANModel as Md
     elif model == "sftgan":
-        from .SFTGAN_ACD_model import SFTGAN_ACD_Model as M
+        from .SFTGAN_ACD import SFTGAN_ACD_Model as Md
     # elif model == 'srragan_hfen':
-    # from .SRRaGAN_hfen_model import SRRaGANModel as M
+    # from .SRRaGAN_hfen_model import SRRaGANModel as Md
     # elif model == 'srragan_n2n':
-    # from .SRRaGAN_n2n_model import SRRaGANModel as M
+    # from .SRRaGAN_n2n_model import SRRaGANModel as Md
     # elif model == 'ESPCN':
-    # from .ESPCN_model import ESPCNModel as M
+    # from .ESPCN_model import ESPCNModel as Md
     elif model == "ppon":
-        from .ppon_model import PPONModel as M
+        from .PPON import PPONModel as Md
     # elif model == 'lpips':
-    # from .LPIPS_model import LPIPSModel as M
+    # from .LPIPS_model import LPIPSModel as Md
     elif model == "asrragan":
-        from .ASRRaGAN_model import ASRRaGANModel as M
+        from .ASRRaGAN import ASRRaGANModel as Md
     else:
-        raise NotImplementedError("Model [{:s}] not recognized.".format(model))
-    m = M(opt)
-    logger.info("Model [{:s}] is created.".format(m.__class__.__name__))
-    return m
+        raise NotImplementedError(f"Model [{model:s}] not recognized.")
+    md = Md(opt)
+    log.info(f"Model [{md.__class__.__name__:s}] is created.")
+    return md
